Now using testify, added a few tests for the home_service.go methods
