+++ openhue/home_service.go
@@ -1,11 +1,10 @@
 package openhue
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	log "github.com/sirupsen/logrus"
 	"openhue-cli/openhue/gen"
+	"slices"
 )
 
 func LoadHome(api *gen.ClientWithResponses) (*Home, error) {
@@ -31,6 +30,90 @@ func LoadHome(api *gen.ClientWithResponses) (*Home, error) {
 	home.Devices = getDevices(ctx, &home.Resource, home.HueData.Children)
 
 	return &home, nil
+}
+
+// FindAllLights returns all the lights that belong to a Home
+func FindAllLights(home *Home) []Light {
+	var lights []Light
+
+	for _, room := range home.Rooms {
+		for _, device := range room.Devices {
+			if device.Light != nil {
+				lights = append(lights, *device.Light)
+			}
+		}
+	}
+
+	return lights
+}
+
+// FindLightByName returns a slice containing a single Light. The slice will be empty if no light was found.
+func FindLightByName(home *Home, name string) []Light {
+	for _, room := range home.Rooms {
+		for _, device := range room.Devices {
+			if device.Light != nil {
+				if device.Light.Name == name {
+					return []Light{*device.Light}
+				}
+			}
+		}
+	}
+
+	log.Warn("no light found with name ", name)
+	return []Light{}
+}
+
+func FindLightsByIds(home *Home, ids []string) []Light {
+
+	var lights []Light
+
+	for _, room := range home.Rooms {
+		for _, device := range room.Devices {
+			if device.Light != nil {
+				if slices.Contains(ids, device.Light.Id) {
+					lights = append(lights, *device.Light)
+				}
+			}
+		}
+	}
+
+	return lights
+}
+
+//
+// Room
+//
+
+func FindAllRooms(home *Home) []Room {
+	var rooms []Room
+
+	for _, room := range home.Rooms {
+		rooms = append(rooms, room)
+	}
+
+	return rooms
+}
+
+func FindRoomById(home *Home, id string) []Room {
+	for _, room := range home.Rooms {
+		if room.Id == id {
+			return []Room{room}
+		}
+	}
+
+	log.Warn("no light found with ID ", id)
+	return []Room{}
+}
+
+func FindRoomByName(home *Home, name string) []Room {
+	for _, room := range home.Rooms {
+		if room.Name == name {
+			return []Room{room}
+		}
+	}
+
+	log.Warn("no room found with name ", name)
+	return []Room{}
 }
 
 //
@@ -156,139 +239,4 @@ func getLight(ctx *hueHomeCtx, parent *Resource, services *[]gen.ResourceIdentif
 	}
 
 	return nil, fmt.Errorf("no 'light' service found for resource %s (%s)", parent.Id, parent.Name)
-}
-
-type hueHomeCtx struct {
-	// api
-	api *gen.ClientWithResponses
-	// context
-	home          *gen.BridgeHomeGet
-	rooms         map[string]gen.RoomGet
-	devices       map[string]gen.DeviceGet
-	lights        map[string]gen.LightGet
-	groupedLights map[string]gen.GroupedLightGet
-}
-
-func loadHueHomeCtx(api *gen.ClientWithResponses) (*hueHomeCtx, error) {
-
-	hueHome, err := fetchBridgeHome(api)
-	if err != nil {
-		return nil, err
-	}
-
-	rooms, err := fetchRooms(api)
-	if err != nil {
-		return nil, err
-	}
-
-	devices, err := fetchDevices(api)
-	if err != nil {
-		return nil, err
-	}
-
-	lights, err := fetchLights(api)
-	if err != nil {
-		return nil, err
-	}
-
-	groupedLights, err := fetchGroupedLights(api)
-	if err != nil {
-		return nil, err
-	}
-
-	return &hueHomeCtx{
-		api:           api,
-		home:          hueHome,
-		rooms:         rooms,
-		devices:       devices,
-		lights:        lights,
-		groupedLights: groupedLights,
-	}, nil
-}
-
-func fetchBridgeHome(api *gen.ClientWithResponses) (*gen.BridgeHomeGet, error) {
-	log.Info("Fetching home...")
-
-	resp, err := api.GetBridgeHomesWithResponse(context.Background())
-	if err != nil {
-		return nil, err
-	}
-
-	if len(*(*resp.JSON200).Data) != 1 {
-		return nil, errors.New("more than 1 home attached to the bridge is not supported yet")
-	}
-
-	return &(*(*resp.JSON200).Data)[0], nil
-}
-
-func fetchDevices(api *gen.ClientWithResponses) (map[string]gen.DeviceGet, error) {
-	log.Info("Fetching devices...")
-
-	resp, err := api.GetDevicesWithResponse(context.Background())
-	if err != nil {
-		return nil, err
-	}
-
-	devices := make(map[string]gen.DeviceGet)
-	hueDevices := (*resp.JSON200).Data
-
-	for _, device := range *hueDevices {
-		devices[*device.Id] = device
-	}
-
-	return devices, err
-}
-
-func fetchRooms(api *gen.ClientWithResponses) (map[string]gen.RoomGet, error) {
-	log.Info("Fetching rooms...")
-
-	resp, err := api.GetRoomsWithResponse(context.Background())
-	if err != nil {
-		return nil, err
-	}
-
-	rooms := make(map[string]gen.RoomGet)
-	hueRooms := (*resp.JSON200).Data
-
-	for _, room := range *hueRooms {
-		rooms[*room.Id] = room
-	}
-
-	return rooms, err
-}
-
-func fetchLights(api *gen.ClientWithResponses) (map[string]gen.LightGet, error) {
-	log.Info("Fetching lights...")
-
-	resp, err := api.GetLightsWithResponse(context.Background())
-	if err != nil {
-		return nil, err
-	}
-
-	lights := make(map[string]gen.LightGet)
-	hueLights := (*resp.JSON200).Data
-
-	for _, light := range *hueLights {
-		lights[*light.Id] = light
-	}
-
-	return lights, nil
-}
-
-func fetchGroupedLights(api *gen.ClientWithResponses) (map[string]gen.GroupedLightGet, error) {
-	log.Info("Fetching grouped lights...")
-
-	resp, err := api.GetGroupedLightsWithResponse(context.Background())
-	if err != nil {
-		return nil, err
-	}
-
-	groupedLights := make(map[string]gen.GroupedLightGet)
-	hueGroupedLights := (*resp.JSON200).Data
-
-	for _, groupedLight := range *hueGroupedLights {
-		groupedLights[*groupedLight.Id] = groupedLight
-	}
-
-	return groupedLights, err
 }
